BREAKING (refactor): 完全重构整理css和js
- 统一CSS class 与 id 变量使用 - 连接符命名
- 完全重写引入 js 与 css 方法，参考sd-webUI
- 分离 js 和 css 到不同文件
