--- assets/html/appearance_switcher.html
@@ -1,6 +1,0 @@
-<div class="switch_checkbox" id="apSwitch">
-    <label class="apSwitch">
-        <input type="checkbox" id="apSwitch_checkbox" data-testid="checkbox" />
-        <span class="apSwitch_span">{label}</span>
-    </label>
-</div>
