+++ modules/overwrites.py
@@ -70,24 +70,3 @@ def postprocess_chat_messages(
             return chat_message
         else:
             raise ValueError(f"Invalid message for Chatbot component: {chat_message}")
-
-with open("./assets/custom.js", "r", encoding="utf-8") as f, \
-    open("./assets/external-scripts.js", "r", encoding="utf-8") as f1:
-    customJS = f.read()
-    externalScripts = f1.read()
-
-
-def reload_javascript():
-    print("Reloading javascript...")
-    js = f'<script>{customJS}</script><script async>{externalScripts}</script>'
-    js += '<script async src="https://cdn.jsdelivr.net/npm/marked/marked.min.js"></script>'
-    js += '<script async type="module" src="http://spin.js.org/spin.umd.js"></script><link type="text/css" href="https://spin.js.org/spin.css" rel="stylesheet" />'
-    def template_response(*args, **kwargs):
-        res = GradioTemplateResponseOriginal(*args, **kwargs)
-        res.body = res.body.replace(b'</html>', f'{js}</html>'.encode("utf8"))
-        res.init_headers()
-        return res
-
-    gr.routes.templates.TemplateResponse = template_response
-
-GradioTemplateResponseOriginal = gr.routes.templates.TemplateResponse
