+++ modules/shared.py
@@ -62,3 +62,4 @@ state = State()
 
 modules_path = os.path.dirname(os.path.realpath(__file__))
 chuanhu_path = os.path.dirname(modules_path)
+assets_path = os.path.join(chuanhu_path, "web_assets")
