+++ web_assets/html/appearance_switcher.html
@@ -1,0 +1,6 @@
+<div class="switch-checkbox" id="apSwitch">
+    <label class="apSwitch">
+        <input type="checkbox" id="apSwitch-checkbox" data-testid="checkbox" />
+        <span class="apSwitch-span">{label}</span>
+    </label>
+</div>
