+++ web_assets/html/update.html
@@ -22,8 +22,8 @@
             <button class="btn-update lg primary svelte-cmf5ev" id="update-button" onclick="bgUpdateChuanhu()">{update_btn}</button>
         </div>
         <div id="close-update-btn" class="btn-update-group hideK">
-            <button class="btn-update lg secondary svelte-cmf5ev" id="update-button" onclick="getUpdateInfo()">{seenew_btn}</button>
-            <button class="btn-update lg primary svelte-cmf5ev" id="cancel-button" onclick="cancelUpdate()">{ok_btn}</button>
+            <button class="btn-update lg secondary svelte-cmf5ev" id="seenew-button" onclick="getUpdateInfo()">{seenew_btn}</button>
+            <button class="btn-update lg primary svelte-cmf5ev" id="ok-button" onclick="cancelUpdate()">{ok_btn}</button>
         </div>
     </div>
 </div>
