+++ web_assets/stylesheet/override-gradio.css
@@ -1,0 +1,67 @@
+
+/* 解决container=False时的错误填充 */
+div.form {
+    background: none !important;
+}
+div.no-container {
+    padding: 10px 0 0 0 !important;
+}
+
+/* gradio的页脚信息 */
+footer {
+    /* display: none !important; */
+    margin-top: .2em !important;
+    font-size: 85%;
+}
+
+/* 覆盖 gradio 丑陋的复制按钮样式 */
+.message pre button[title="copy"] {
+    border-radius: 5px;
+    transition: background-color .2s ease;
+}
+.message pre button[title="copy"]:hover {
+    background-color: #333232;
+}
+.message pre button .check {
+    color: #fff !important;
+    background: var(--neutral-950) !important;
+}
+
+
+
+
+/* Override Slider Styles (for webkit browsers like Safari and Chrome)
+ * 好希望这份提案能早日实现 https://github.com/w3c/csswg-drafts/issues/4410
+ * 进度滑块在各个平台还是太不统一了
+**/
+
+input[type="range"] {
+    /* -webkit-appearance: none; */
+    appearance: none;
+    height: 4px;
+    background: var(--input-background-fill);
+    border-radius: 5px;
+    background-image: linear-gradient(var(--primary-500),var(--primary-500));
+    background-size: 0% 100%;
+    background-repeat: no-repeat;
+}
+input[type="range"]::-webkit-slider-thumb {
+    -webkit-appearance: none;
+    height: 20px;
+    width: 20px;
+    border-radius: 50%;
+    border: solid 0.5px #ddd;
+    background-color: white;
+    cursor: ew-resize;
+    box-shadow: var(--input-shadow);
+    transition: background-color .1s ease;
+}
+input[type="range"]::-webkit-slider-thumb:hover {
+    background: var(--neutral-50);
+}
+input[type=range]::-webkit-slider-runnable-track {
+    -webkit-appearance: none;
+    box-shadow: none;
+    border: none;
+    background: transparent;
+}
